requirements.txt
beautifulsoup4==4.12.2
numpy==1.26.0
pandas==2.1.1
pybase64==1.3.1
python-dateutil==2.8.2
python-magic==0.4.27
pytz==2023.3.post1
six==1.16.0
soupsieve==2.5
tqdm==4.66.1
tzdata==2023.3
